backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
asyncpg==0.29.0  # PostgreSQL async driver
aiosqlite==0.19.0  # SQLite async driver
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9
aiofiles==23.2.1
psycopg2-binary==2.9.9  # For PostgreSQL support on PythonAnywhere 